1398 - Moves object serializer to utils and adds unit tests
